Refactor drop zone boundary
