+++ src/pages/Search/SearchPage.tsx
@@ -1,42 +1,33 @@
 import React, {useCallback, useEffect, useMemo, useRef, useState} from 'react';
 import Animated from 'react-native-reanimated';
-import DragAndDropConsumer from '@components/DragAndDrop/Consumer';
-import DragAndDropProvider from '@components/DragAndDrop/Provider';
-import DropZoneUI from '@components/DropZone/DropZoneUI';
 import {useSearchActionsContext, useSearchStateContext} from '@components/Search/SearchContext';
 import type {SearchParams} from '@components/Search/types';
 import {usePlaybackActionsContext} from '@components/VideoPlayerContexts/PlaybackContext';
 import useConfirmReadyToOpenApp from '@hooks/useConfirmReadyToOpenApp';
 import useFilterFormValues from '@hooks/useFilterFormValues';
-import {useMemoizedLazyExpensifyIcons} from '@hooks/useLazyAsset';
-import useLocalize from '@hooks/useLocalize';
 import useMobileSelectionMode from '@hooks/useMobileSelectionMode';
 import usePrevious from '@hooks/usePrevious';
-import useReceiptScanDrop from '@hooks/useReceiptScanDrop';
 import useResponsiveLayout from '@hooks/useResponsiveLayout';
 import useSearchFilterSync from '@hooks/useSearchFilterSync';
 import useSearchShouldCalculateTotals from '@hooks/useSearchShouldCalculateTotals';
-import useTheme from '@hooks/useTheme';
 import useThemeStyles from '@hooks/useThemeStyles';
 import {searchInServer} from '@libs/actions/Report';
 import {search} from '@libs/actions/Search';
 import type {PlatformStackScreenProps} from '@libs/Navigation/PlatformStackNavigation/types';
 import type {SearchFullscreenNavigatorParamList} from '@libs/Navigation/types';
 import {buildSearchQueryJSON} from '@libs/SearchQueryUtils';
-import variables from '@styles/variables';
 import CONST from '@src/CONST';
 import type SCREENS from '@src/SCREENS';
 import type {SearchResults} from '@src/types/onyx';
 import SearchPageNarrow from './SearchPageNarrow';
 import SearchPageWide from './SearchPageWide';
+import SearchReceiptDropBoundary from './SearchReceiptDropBoundary';
 
 type SearchPageProps = PlatformStackScreenProps<SearchFullscreenNavigatorParamList, typeof SCREENS.SEARCH.ROOT>;
 
 function SearchPage({route}: SearchPageProps) {
-    const {translate} = useLocalize();
     const {shouldUseNarrowLayout} = useResponsiveLayout();
     const styles = useThemeStyles();
-    const theme = useTheme();
     const {selectedTransactions, lastSearchType, areAllMatchingItemsSelected, currentSearchKey, currentSearchResults} = useSearchStateContext();
     const {clearSelectedTransactions, setLastSearchType} = useSearchActionsContext();
     const isMobileSelectionModeEnabled = useMobileSelectionMode(clearSelectedTransactions);
@@ -44,7 +35,6 @@ function SearchPage({route}: SearchPageProps) {
     const queryJSON = useMemo(() => buildSearchQueryJSON(route.params.q, route.params.rawQuery), [route.params.q, route.params.rawQuery]);
     const filterFormValues = useFilterFormValues(queryJSON);
     useSearchFilterSync(filterFormValues);
-    const expensifyIcons = useMemoizedLazyExpensifyIcons(['SmartScan'] as const);
 
     const lastNonEmptySearchResults = useRef<SearchResults | undefined>(undefined);
 
@@ -62,8 +52,6 @@ function SearchPage({route}: SearchPageProps) {
     }, [lastSearchType, queryJSON, setLastSearchType, currentSearchResults]);
 
     const selectedTransactionsKeys = Object.keys(selectedTransactions ?? {});
-
-    const {initScanRequest, PDFValidationComponent, ErrorModal, isDragDisabled} = useReceiptScanDrop();
     const {resetVideoPlayerData} = usePlaybackActionsContext();
 
     const [isSorting, setIsSorting] = useState(false);
@@ -141,31 +129,17 @@ function SearchPage({route}: SearchPageProps) {
         setIsSorting(true);
     }, []);
 
-    return (
-        <Animated.View style={[styles.flex1]}>
-            {shouldUseNarrowLayout ? (
-                <DragAndDropProvider isDisabled={isDragDisabled}>
-                    {PDFValidationComponent}
-                    <SearchPageNarrow
-                        queryJSON={queryJSON}
-                        metadata={metadata}
-                        searchResults={searchResults}
-                        isMobileSelectionModeEnabled={isMobileSelectionModeEnabled}
-                        footerData={footerData}
-                        shouldShowFooter={shouldShowFooter}
-                    />
-                    <DragAndDropConsumer onDrop={initScanRequest}>
-                        <DropZoneUI
-                            icon={expensifyIcons.SmartScan}
-                            dropTitle={translate('dropzone.scanReceipts')}
-                            dropStyles={styles.receiptDropOverlay(true)}
-                            dropTextStyles={styles.receiptDropText}
-                            dropWrapperStyles={{marginBottom: variables.bottomTabHeight}}
-                            dashedBorderStyles={[styles.dropzoneArea, styles.easeInOpacityTransition, styles.activeDropzoneDashedBorder(theme.receiptDropBorderColorActive, true)]}
-                        />
-                    </DragAndDropConsumer>
-                    {ErrorModal}
-                </DragAndDropProvider>
+    const pageContent = useMemo(
+        () =>
+            shouldUseNarrowLayout ? (
+                <SearchPageNarrow
+                    queryJSON={queryJSON}
+                    metadata={metadata}
+                    searchResults={searchResults}
+                    isMobileSelectionModeEnabled={isMobileSelectionModeEnabled}
+                    footerData={footerData}
+                    shouldShowFooter={shouldShowFooter}
+                />
             ) : (
                 <SearchPageWide
                     queryJSON={queryJSON}
@@ -176,13 +150,27 @@ function SearchPage({route}: SearchPageProps) {
                     handleSearchAction={handleSearchAction}
                     onSortPressedCallback={onSortPressedCallback}
                     route={route}
-                    initScanRequest={initScanRequest}
-                    isDragDisabled={isDragDisabled}
-                    PDFValidationComponent={PDFValidationComponent}
-                    ErrorModal={ErrorModal}
                     shouldShowFooter={shouldShowFooter}
                 />
-            )}
+            ),
+        [
+            shouldUseNarrowLayout,
+            queryJSON,
+            metadata,
+            searchResults,
+            isMobileSelectionModeEnabled,
+            footerData,
+            shouldShowFooter,
+            searchRequestResponseStatusCode,
+            handleSearchAction,
+            onSortPressedCallback,
+            route,
+        ],
+    );
+
+    return (
+        <Animated.View style={[styles.flex1]}>
+            <SearchReceiptDropBoundary shouldUseNarrowLayout={shouldUseNarrowLayout}>{pageContent}</SearchReceiptDropBoundary>
         </Animated.View>
     );
 }
