+++ src/pages/Search/SearchPageWide.tsx
@@ -3,9 +3,6 @@ import type {NativeScrollEvent, NativeSyntheticEvent} from 'react-native';
 import {View} from 'react-native';
 import type {OnyxEntry} from 'react-native-onyx';
 import FullPageNotFoundView from '@components/BlockingViews/FullPageNotFoundView';
-import DragAndDropConsumer from '@components/DragAndDrop/Consumer';
-import DragAndDropProvider from '@components/DragAndDrop/Provider';
-import DropZoneUI from '@components/DropZone/DropZoneUI';
 import ScreenWrapper from '@components/ScreenWrapper';
 import {ScrollOffsetContext} from '@components/ScrollOffsetContextProvider';
 import Search from '@components/Search';
@@ -13,9 +10,6 @@ import SearchPageFooter from '@components/Search/SearchPageFooter';
 import SearchFiltersBar from '@components/Search/SearchPageHeader/SearchFiltersBar';
 import SearchPageHeader from '@components/Search/SearchPageHeader/SearchPageHeader';
 import type {SearchParams, SearchQueryJSON} from '@components/Search/types';
-import {useMemoizedLazyExpensifyIcons} from '@hooks/useLazyAsset';
-import useLocalize from '@hooks/useLocalize';
-import useTheme from '@hooks/useTheme';
 import useThemeStyles from '@hooks/useThemeStyles';
 import type {PlatformStackRouteProp} from '@libs/Navigation/PlatformStackNavigation/types';
 import type {SearchFullscreenNavigatorParamList} from '@libs/Navigation/types';
@@ -38,10 +32,6 @@ type SearchPageWideProps = {
     handleSearchAction: (value: SearchParams | string) => void;
     onSortPressedCallback: () => void;
     route: PlatformStackRouteProp<SearchFullscreenNavigatorParamList, typeof SCREENS.SEARCH.ROOT>;
-    initScanRequest: (e: DragEvent) => void;
-    isDragDisabled: boolean;
-    PDFValidationComponent: React.ReactNode;
-    ErrorModal: React.ReactNode;
     shouldShowFooter: boolean;
 };
 
@@ -54,15 +44,9 @@ function SearchPageWide({
     handleSearchAction,
     onSortPressedCallback,
     route,
-    initScanRequest,
-    isDragDisabled,
-    PDFValidationComponent,
-    ErrorModal,
     shouldShowFooter,
 }: SearchPageWideProps) {
     const styles = useThemeStyles();
-    const theme = useTheme();
-    const {translate} = useLocalize();
     const {saveScrollOffset} = useContext(ScrollOffsetContext);
 
     const scrollHandler = useCallback(
@@ -84,7 +68,6 @@ function SearchPageWide({
         return [styles.mtAuto];
     }, [shouldShowFooter, styles]);
 
-    const expensifyIcons = useMemoizedLazyExpensifyIcons(['SmartScan']);
     const handleOnBackButtonPress = () => Navigation.goBack(ROUTES.SEARCH_ROOT.getRoute({query: buildCannedSearchQuery()}));
 
     return (
@@ -102,8 +85,7 @@ function SearchPageWide({
                     shouldShowLink={false}
                 >
                     {!!queryJSON && (
-                        <DragAndDropProvider isDisabled={isDragDisabled}>
-                            {PDFValidationComponent}
+                        <>
                             <SearchPageHeader
                                 queryJSON={queryJSON}
                                 handleSearch={handleSearchAction}
@@ -130,20 +112,10 @@ function SearchPageWide({
                                     currency={footerData.currency}
                                 />
                             )}
-                            <DragAndDropConsumer onDrop={initScanRequest}>
-                                <DropZoneUI
-                                    icon={expensifyIcons.SmartScan}
-                                    dropTitle={translate('dropzone.scanReceipts')}
-                                    dropStyles={styles.receiptDropOverlay(true)}
-                                    dropTextStyles={styles.receiptDropText}
-                                    dashedBorderStyles={[styles.dropzoneArea, styles.easeInOpacityTransition, styles.activeDropzoneDashedBorder(theme.receiptDropBorderColorActive, true)]}
-                                />
-                            </DragAndDropConsumer>
-                        </DragAndDropProvider>
+                        </>
                     )}
                 </FullPageNotFoundView>
             </ScreenWrapper>
-            {ErrorModal}
         </View>
     );
 }
